ECO-269: refactor VestingChart to let class VestingDetail computed points by itself.
